[Pal] Move x86_64-specific CPUID into cpu.h, pal-arch.h, db_main-x86_64.c
