+++ Documentation/Doxyfile-pal
@@ -5,4 +5,4 @@ INPUT = ../Pal/include/arch/x86_64 ../Pal/include/pal ../Pal/src
 EXCLUDE = ../Pal/src/host
 MACRO_EXPANSION = YES
 EXPAND_ONLY_PREDEF = YES
-PREDEFINED = noreturn=
+PREDEFINED = noreturn= __x86_64__=1
