+++ Pal/src/host/Linux-SGX/Makefile
@@ -37,6 +37,7 @@ enclave-objs = \
 	db_exception.o \
 	db_files.o \
 	db_main.o \
+	db_main-$(ARCH).o \
 	db_memory.o \
 	db_misc.o \
 	db_mutex.o \
