+++ Pal/src/host/Linux/Makefile
@@ -36,6 +36,7 @@ objs = \
 	db_exception.o \
 	db_files.o \
 	db_main.o \
+	db_main-$(ARCH).o \
 	db_memory.o \
 	db_misc.o \
 	db_mutex.o \
