+++ Pal/src/host/Linux/db_main.c
@@ -9,7 +9,6 @@
  */
 
 #include "api.h"
-#include "bogomips.h"
 #include "pal.h"
 #include "pal_debug.h"
 #include "pal_defs.h"
@@ -271,74 +270,11 @@ void pal_linux_main(void* initial_rsp, void* fini_callback) {
              first_process ? argv + 2 : argv + 3, envp);
 }
 
-/* the following code is borrowed from CPUID */
-void cpuid (unsigned int leaf, unsigned int subleaf,
-            unsigned int words[])
-{
-  __asm__ ("cpuid"
-      : "=a" (words[PAL_CPUID_WORD_EAX]),
-        "=b" (words[PAL_CPUID_WORD_EBX]),
-        "=c" (words[PAL_CPUID_WORD_ECX]),
-        "=d" (words[PAL_CPUID_WORD_EDX])
-      : "a" (leaf),
-        "c" (subleaf));
-}
-
-#define FOUR_CHARS_VALUE(s, w)      \
-    (s)[0] = (w) & 0xff;            \
-    (s)[1] = ((w) >>  8) & 0xff;    \
-    (s)[2] = ((w) >> 16) & 0xff;    \
-    (s)[3] = ((w) >> 24) & 0xff;
-
-#define BPI  32
-#define POWER2(power) \
-   (1ULL << (power))
-#define RIGHTMASK(width) \
-   (((unsigned long)(width) >= BPI) ? ~0ULL : POWER2(width) - 1ULL)
-
-#define BIT_EXTRACT_LE(value, start, after) \
-   (((unsigned long)(value) & RIGHTMASK(after)) >> start)
-
-static char * cpu_flags[]
-      = { "fpu",    // "x87 FPU on chip"
-          "vme",    // "virtual-8086 mode enhancement"
-          "de",     // "debugging extensions"
-          "pse",    // "page size extensions"
-          "tsc",    // "time stamp counter"
-          "msr",    // "RDMSR and WRMSR support"
-          "pae",    // "physical address extensions"
-          "mce",    // "machine check exception"
-          "cx8",    // "CMPXCHG8B inst."
-          "apic",   // "APIC on chip"
-          NULL,
-          "sep",    // "SYSENTER and SYSEXIT"
-          "mtrr",   // "memory type range registers"
-          "pge",    // "PTE global bit"
-          "mca",    // "machine check architecture"
-          "cmov",   // "conditional move/compare instruction"
-          "pat",    // "page attribute table"
-          "pse36",  // "page size extension"
-          "pn",     // "processor serial number"
-          "clflush",    // "CLFLUSH instruction"
-          NULL,
-          "dts",    // "debug store"
-          "acpi",   // "Onboard thermal control"
-          "mmx",    // "MMX Technology"
-          "fxsr",   // "FXSAVE/FXRSTOR"
-          "sse",    // "SSE extensions"
-          "sse2",   // "SSE2 extensions"
-          "ss",     // "self snoop"
-          "ht",     // "hyper-threading / multi-core supported"
-          "tm",     // "therm. monitor"
-          "ia64",   // "IA64"
-          "pbe",    // "pending break event"
-        };
-
 /*
  * Returns the number of online CPUs read from /sys/devices/system/cpu/online, -errno on failure.
  * Understands complex formats like "1,3-5,6".
  */
-static int get_cpu_count(void) {
+int get_cpu_count(void) {
     int fd = INLINE_SYSCALL(open, 3, "/sys/devices/system/cpu/online", O_RDONLY|O_CLOEXEC, 0);
     if (fd < 0)
         return unix_to_pal_error(ERRNO(fd));
@@ -381,7 +317,7 @@ static int get_cpu_count(void) {
     return cpu_count;
 }
 
-static ssize_t read_file_buffer(const char* filename, char* buf, size_t buf_size) {
+ssize_t read_file_buffer(const char* filename, char* buf, size_t buf_size) {
     int fd;
 
     fd = INLINE_SYSCALL(open, 2, filename, O_RDONLY);
@@ -392,96 +328,4 @@ static ssize_t read_file_buffer(const char* filename, char* buf, size_t buf_size
     INLINE_SYSCALL(close, 1, fd);
 
     return n;
-}
-
-static double get_bogomips(void) {
-    char buf[2048];
-    ssize_t len;
-
-    len = read_file_buffer("/proc/cpuinfo", buf, sizeof(buf) - 1);
-    if (len < 0)
-        return 0.0;
-    buf[len] = 0;
-
-    return sanitize_bogomips_value(get_bogomips_from_cpuinfo_buf(buf));
-}
-
-int _DkGetCPUInfo (PAL_CPU_INFO * ci)
-{
-    unsigned int words[PAL_CPUID_WORD_NUM];
-    int rv = 0;
-
-    const size_t VENDOR_ID_SIZE = 13;
-    char* vendor_id = malloc(VENDOR_ID_SIZE);
-    cpuid(0, 0, words);
-
-    FOUR_CHARS_VALUE(&vendor_id[0], words[PAL_CPUID_WORD_EBX]);
-    FOUR_CHARS_VALUE(&vendor_id[4], words[PAL_CPUID_WORD_EDX]);
-    FOUR_CHARS_VALUE(&vendor_id[8], words[PAL_CPUID_WORD_ECX]);
-    vendor_id[VENDOR_ID_SIZE - 1] = '\0';
-    ci->cpu_vendor = vendor_id;
-
-    const size_t BRAND_SIZE = 49;
-    char* brand = malloc(BRAND_SIZE);
-    cpuid(0x80000002, 0, words);
-    memcpy(&brand[ 0], words, sizeof(unsigned int) * PAL_CPUID_WORD_NUM);
-    cpuid(0x80000003, 0, words);
-    memcpy(&brand[16], words, sizeof(unsigned int) * PAL_CPUID_WORD_NUM);
-    cpuid(0x80000004, 0, words);
-    memcpy(&brand[32], words, sizeof(unsigned int) * PAL_CPUID_WORD_NUM);
-    brand[BRAND_SIZE - 1] = '\0';
-    ci->cpu_brand = brand;
-
-    /* we cannot use CPUID(0xb) because it counts even disabled-by-BIOS cores (e.g. HT cores);
-     * instead we extract info on number of online CPUs by parsing sysfs pseudo-files */
-    int cores = get_cpu_count();
-    if (cores < 0) {
-        free(vendor_id);
-        free(brand);
-        return cores;
-    }
-    ci->cpu_num = cores;
-
-    cpuid(1, 0, words);
-    ci->cpu_family   = BIT_EXTRACT_LE(words[PAL_CPUID_WORD_EAX],  8, 12);
-    ci->cpu_model    = BIT_EXTRACT_LE(words[PAL_CPUID_WORD_EAX],  4,  8);
-    ci->cpu_stepping = BIT_EXTRACT_LE(words[PAL_CPUID_WORD_EAX],  0,  4);
-
-    if (!memcmp(vendor_id, "GenuineIntel", 12) ||
-        !memcmp(vendor_id, "AuthenticAMD", 12)) {
-        ci->cpu_family += BIT_EXTRACT_LE(words[PAL_CPUID_WORD_EAX], 20, 28);
-        ci->cpu_model  += BIT_EXTRACT_LE(words[PAL_CPUID_WORD_EAX], 16, 20) << 4;
-    }
-
-    int flen = 0, fmax = 80;
-    char * flags = malloc(fmax);
-
-    for (int i = 0 ; i < 32 ; i++) {
-        if (!cpu_flags[i])
-            continue;
-
-        if (BIT_EXTRACT_LE(words[PAL_CPUID_WORD_EDX], i, i + 1)) {
-            int len = strlen(cpu_flags[i]);
-            if (flen + len + 1 > fmax) {
-                char * new_flags = malloc(fmax * 2);
-                memcpy(new_flags, flags, flen);
-                free(flags);
-                fmax *= 2;
-                flags = new_flags;
-            }
-            memcpy(flags + flen, cpu_flags[i], len);
-            flen += len;
-            flags[flen++] = ' ';
-        }
-    }
-
-    flags[flen ? flen - 1 : 0] = 0;
-    ci->cpu_flags = flags;
-
-    ci->cpu_bogomips = get_bogomips();
-    if (ci->cpu_bogomips == 0.0) {
-        printf("Warning: bogomips could not be retrieved, passing 0.0 to the application\n");
-    }
-
-    return rv;
 }
