+++ Pal/src/host/Linux/pal_linux.h
@@ -132,6 +132,9 @@ bool stataccess (struct stat * stats, int acc);
 void init_child_process(int parent_pipe_fd, PAL_HANDLE* parent, PAL_HANDLE* exec,
                         PAL_HANDLE* manifest);
 
+int get_cpu_count(void);
+ssize_t read_file_buffer(const char* filename, char* buf, size_t buf_size);
+
 void cpuid (unsigned int leaf, unsigned int subleaf,
             unsigned int words[]);
 int block_async_signals (bool block);
